REFACTORING TOWARDS AN OBJECT-ORIENTED DESIGN

+++ src/main/java/com/morrisco/net/eCommerceSystem/entities/CartItem.java
@@ -29,9 +29,9 @@ public class CartItem {
     private Product product;
 
     @Column(name = "quantity")
-    private Integer quantity;
+    private BigDecimal quantity;
 
     public  BigDecimal getTotalPrice(){
-        return product.getPrice().multiply(BigDecimal.valueOf(quantity));
+        return product.getPrice().multiply(quantity);
     }
 }
